labs/lab06/requirements.txt
paho-mqtt
lgpio